fix(gcs): harden clear-sign parsing and canonical validation
- reject unsupported visibility constraints instead of silently ignoring them
- validate canonical uint, address and selector encodings across GCS formatters
- prevent token arrays from reusing metadata from previous elements
- continue validating trusted-name arrays after filtered elements
- reject zero-length and oversized numeric values
- require root and nested ABI calldata to be 32-byte aligned
- prevent NULL pointer arithmetic when handling empty values
- validate unit and threshold display strings as printable text
- use no-throw hashing for transaction metadata parsing
- remove redundant proxy cleanup from the GCS UI lifecycle
- add fuzz regression guards for canonical conversion boundaries

Tests:
- 5/5 unit tests passed
- ASan/UBSan corpus replay passed for GCS, TIP-712, external
  metadata, INS_SIGN and TriggerSmartContract decoding

+++ src/tron-sdk/common_utils.c
@@ -242,7 +242,7 @@ bool adjustDecimals(const char *src,
 
 // change "4.2 ENS" to "ENS 4.2"
 bool amountToString(const uint8_t *amount,
-                    uint8_t amount_size,
+                    size_t amount_size,
                     uint8_t decimals,
                     const char *ticker,
                     char *out_buffer,
